pyproject.toml
[tool.poetry]
name = "zaber_slit"
version = "0.1.0"
description = ""
authors = ["Zach Mitchell <zmitchell@fastmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
pyserial = "^3.4"
click = "^7.1.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry.scripts]
slit = "zaber_slit.main:cli"